UPDATE Logon Process
Return value for authentication using the version 10 api deffers from version 9.
Version 9 returns the token as the value of the CyberArkLogonResult property.
Version 10 returns the raw value with no property.

Logon process updated to include the additional parameters and URL endpoint for version 10.
If Invoke-PASRestMethod is called from New-PASSession, and the raw value is returned (i.e. its a version 10 authentication request), the return value will be added to the CyberArkLogonResult property.

Additional exception handling added to cater for failures related to v10 endpoint being used with a v9 solution.