ADD Assert-VersionRequirement
Added function which can be used to throw an error and prevent a script being run if an attempt is made to invoke a function against a version of CyberArk which does not support the API call, or specific parameters of an API method.

Relates to #70

Module functions are required to be updated to call Assert-VersionRequirement.