ADD Get-ByteArray
Replaced code t read all bytes of a file to a byte array with call to Get-ByteArray function.
Explicitly set debug to false on call to Invoke-PASRestMethod in functions that send files as byte arrays to the API - debug output of the contents of the byte array breaks the upload and causes a server 500 error to be returned.