FIX Get-PASAccountPassword
Fixes bug. CyberArk version 9 provides account password as a Byte Array. Version 10 provides a string.
An update to the module for version 10 compatibility broke backwards compatibility for version 9. This commit fixes that.

Added code to Get-PASAccountPassword which checks the object type name of the returned data indicates a Byte Array has been returned, the Byte Array is converted to a string.

Added Pester tests to check the condition for the return types expected from each version.

relates to #92