:recycle: UPDATE Request Header
Request header is now provided via Script Scope WebSession Object.
Updated functions which require additional header elements to create a copy of the $Script:WebSession variable in the local scope which then has the required Header elements added to it.
The local scope WebSession is then sent with the request.