:recycle: :fire: UPDATE New-PASSession
Removed parameter `$SessionVariable`.
SessionVariable will always be sent with a logon request. Replaced parameter with hardcoded value; with the recent updates to the module there is no longer value in allowing an arbitary value to be set for the value of the SessionVariable parameter.
`psPAS` functionality is unaffected. `New-PASSession` is simplified.