:heavy_plus_sign: ADD Get-PASResponse
Initial commit of `Get-PASResponse`.
Intended to be used to abstract the logic currently contained in `Invoke-PASRestMethod` which returns the content of the web response recieved from the API.

Logic has been ammended to always pass the content by default (so there is no longer a need to process responses of content type octet-stream & application/save separatley.
Content-type, along with the name of the initial function can be used to determine special processing of the result as required.