:zap: :sparkles: UPDATE File Output
Moving logic to save byte stream output to file to `Out-PASFile`.
Updated `Get-PASPSMConnectionParameter` to output an RDP file by default, or the HTML5 GW details.
`Get-PASResponse` updated to return api response content and header if content type is `octet-stream`. This allows `Out-PASFile` to use information in the response header to name the output file.