Added -RequiresPVWAIISAuthentication & OTP prompt
Added -RequiresPVWAIISAuthentication to support the usecase where PVWA web access requires windows authentication before any login method can be used (except CyberArk login).
Some OTP solutions send the token via SMS. This means that the token is not known the moment the request is sent. Input prompt is needed to enter the token and a new webservice request to send it to CyberArk.