:boom: :recycle: UPDATE Get-PASSafeMember
Changes output of function to enable permission settings with non boolean values to be returned.
Previous output for "Permissions" was an array of the permission names which are granted to a user or group on a safe.
Updated output for "Permissions" is an object containing the property name of each permission with a true/false/integer value.
#275