:recycle: :alien: UPDATE Get-PASPTAEvent
Between versions 11.2 & 11.3, the documentation for the PTA Events endpoint changes; `lastUpdatedEventDate` moves from being specified in the request header to in the request url.
Understanding why events were not being filtered as expected reveals that the correct parameter to include in the url is actually `fromUpdateDate`, this also appears to be a unixtimestamp in milleseconds instead of seconds.

- `Get-PASPTAEvent`
  - Adds parameter `fromUpdateDate`.
    - this is converted into a unixtimestamp in millesends for inclusion int he request url
  - Removes parameter `UseLegacyMethod`.
    - No longer required as specifying a value for `lastUpdatedEventDate` will now always include this in the request header.
  - Lowers required version from 11.4 to 11.3 when using `fromUpdateDate` & `status` parameter combinations.
    - Assume `fromUpdateDate` was introduced in 11.3
    - `status` is documented from 11.3
- `ConvertTo-UnixTime`
  - Adds `Milliseconds` option, for when value is required in milliesends instead of seconds.