:ambulance: :recycle: FIX Set-PASAccount
Fixes condition where attempting updates to multiple properties of an account results in an error.
ParameterSetName was recently changed from `V10MultiOp` to `Gen2MultiOp`, but the logic in the function still referenced the old name.
Updating code to reflect correct ParameterSetName `Gen2MultiOp`, this resuts in JSON payload being formatted as required.
Removes  `-Depth 3` parameter from `ConvertTo-Json`; the correctly formatted JSON does not have a depth of 3, so this is not required.
#326