UPDATE New-PASSession
Adds logic around getting the logged on user name.
Current logged on user details are not able to be queried in recent versions of privilege cloud.
If the username is not able to be retrieved, the username from the used credential object is used instead. For other scenarios where the credential object cannot be used, the username is returned as a null value.