UPDATE Get-PASSession
Adds `LastError` & `LastErrorTime` to output reported by `Get-PASSession`.
Whenever an error is encountered executing a psPAS command, it will be stored in the script scope object for easy retrieval and review in case correct parsing of the error is not possible or the error is reported clearly in the console.