Handle Negative Expiry Dates
If for some reason the safe member being updated has a negative expirydate (less than zero in unix time), and the expirydate is not explicitly being updated in the set command - ensure the value sent for the expirydate in the request is null.

Avoids a condition where an existing negative expiry date on a safe member causes an out of bound exception on safe member update.

Updates test file to check for described condition